Replace config.json with environment variables
Removed the usage of a configuration file (config.json) in favor of using environment variables. This change improves security by allowing sensitive data (such as API keys) to be stored in hidden .env files, which are now included in the .gitignore file to prevent accidental upload. The envGet function has been implemented to fetch these variables.