Implement responsive table design for DNS Element
The updates in the code make the DNS Element table more responsive and suitable for viewing in different screen sizes by hiding certain columns for smaller devices. This change improves the usability of the interface and makes it more accessible across a multitude of devices. Moreover, the 'dev' command is added to run configurations for easier application development.